# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Numerology CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loshu/missing:
    get:
      operationId: listMissingNumberRules
      tags: [loshu]
      summary: List missing number rules
      responses:
        "200":
          description: Missing number rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MissingNumberRule"
    post:
      operationId: createMissingNumberRule
      tags: [loshu]
      summary: Create a missing number rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissingNumberRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingNumberRule"
  /loshu/missing/{id}:
    patch:
      operationId: updateMissingNumberRule
      tags: [loshu]
      summary: Update a missing number rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissingNumberRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingNumberRule"
    delete:
      operationId: deleteMissingNumberRule
      tags: [loshu]
      summary: Delete a missing number rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /loshu/repeated:
    get:
      operationId: listRepeatedNumberRules
      tags: [loshu]
      summary: List repeated number rules
      responses:
        "200":
          description: Repeated number rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepeatedNumberRule"
    post:
      operationId: createRepeatedNumberRule
      tags: [loshu]
      summary: Create a repeated number rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepeatedNumberRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepeatedNumberRule"
  /loshu/repeated/{id}:
    patch:
      operationId: updateRepeatedNumberRule
      tags: [loshu]
      summary: Update a repeated number rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepeatedNumberRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepeatedNumberRule"
    delete:
      operationId: deleteRepeatedNumberRule
      tags: [loshu]
      summary: Delete a repeated number rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /loshu/arrows:
    get:
      operationId: listArrowRules
      tags: [loshu]
      summary: List Lo Shu arrow rules
      responses:
        "200":
          description: Arrow rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArrowRule"
    post:
      operationId: createArrowRule
      tags: [loshu]
      summary: Create an arrow rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArrowRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrowRule"
  /loshu/arrows/{id}:
    patch:
      operationId: updateArrowRule
      tags: [loshu]
      summary: Update an arrow rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArrowRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArrowRule"
    delete:
      operationId: deleteArrowRule
      tags: [loshu]
      summary: Delete an arrow rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Reports ──────────────────────────────────────────────────────────────────
  /reports/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a numerology report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumerologyReport"
components:
  schemas:
...
    MissingNumberRule:
      type: object
      properties:
        id:
          type: integer
        missing_number:
          type: integer
        title:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - missing_number
        - title
        - strengths
        - weaknesses
        - recommendations
        - is_active
        - created_at
        - updated_at
    MissingNumberRuleInput:
      type: object
      properties:
        missing_number:
          type: integer
        title:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        is_active:
          type: boolean
      required:
        - missing_number
        - title
        - strengths
        - weaknesses
        - recommendations
    MissingNumberRuleUpdate:
      type: object
      properties:
        missing_number:
          type: integer
        title:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        is_active:
          type: boolean
    RepeatedNumberRule:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        count:
          type: integer
        title:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - number
        - count
        - title
        - strengths
        - weaknesses
        - is_active
        - created_at
        - updated_at
    RepeatedNumberRuleInput:
      type: object
      properties:
        number:
          type: integer
        count:
          type: integer
        title:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        is_active:
          type: boolean
      required:
        - number
        - count
        - title
        - strengths
        - weaknesses
    RepeatedNumberRuleUpdate:
      type: object
      properties:
        number:
          type: integer
        count:
          type: integer
        title:
          type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        is_active:
          type: boolean
    ArrowRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        numbers:
          type: array
          items:
            type: integer
        interpretation:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
      required:
        - id
        - name
        - numbers
        - interpretation
        - is_active
        - created_at
        - updated_at
    ArrowRuleInput:
      type: object
      properties:
        name:
          type: string
        numbers:
          type: array
          items:
            type: integer
        interpretation:
          type: string
        is_active:
          type: boolean
      required:
        - name
        - numbers
        - interpretation
    ArrowRuleUpdate:
      type: object
      properties:
        name:
          type: string
        numbers:
          type: array
          items:
            type: integer
        interpretation:
          type: string
        is_active:
          type: boolean
    ReportRequest:
      type: object
      properties:
        name:
          type: string
        date_of_birth:
          type: string
          description: "ISO date string: YYYY-MM-DD"
      required:
        - name
        - date_of_birth
    NumerologyReport:
      type: object
      properties:
        subject:
          type: object
          properties:
            name:
              type: string
            date_of_birth:
              type: string
          required:
            - name
            - date_of_birth
        numbers:
          $ref: "#/components/schemas/CalculatedNumbers"
        interpretations:
          type: array
          items:
            $ref: "#/components/schemas/RuleInterpretation"
        lo_shu:
          $ref: "#/components/schemas/LoShuResult"
        generated_at:
          type: string
      required:
        - subject
        - numbers
        - interpretations
        - lo_shu
        - generated_at
    CalculatedNumbers:
      type: object
      properties:
        birthday_number:
          type: integer
        destiny_number:
          type: integer
        personality_number:
          type: ["integer", "null"]
        personal_year:
          type: integer
        personal_month:
          type: integer
        personal_day:
          type: integer
      required:
        - birthday_number
        - destiny_number
        - personal_year
        - personal_month
        - personal_day
    RuleInterpretation:
      type: object
      properties:
        rule_id:
          type: integer
        rule_name:
          type: string
        rule_type:
          type: string
        keywords:
          type: array
          items:
            type: string
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
      required:
        - rule_id
        - rule_name
        - rule_type
        - keywords
        - strengths
        - weaknesses
        - recommendations
    LoShuResult:
      type: object
      properties:
        grid:
          type: array
          items:
            type: array
            items:
              type: integer
        missing_numbers:
          type: array
          items:
            type: integer
        repeated_numbers:
          type: array
          items:
            $ref: "#/components/schemas/RepeatedEntry"
        active_arrows:
          type: array
          items:
            type: string
        missing_interpretations:
          type: array
          items:
            $ref: "#/components/schemas/MissingNumberRule"
        repeated_interpretations:
          type: array
          items:
            $ref: "#/components/schemas/RepeatedNumberRule"
        arrow_interpretations:
          type: array
          items:
            $ref: "#/components/schemas/ArrowRule"
      required:
        - grid
        - missing_numbers
        - repeated_numbers
        - active_arrows
        - missing_interpretations
        - repeated_interpretations
        - arrow_interpretations
    RepeatedEntry:
      type: object
      properties:
        number:
          type: integer
        count:
          type: integer
      required:
        - number
        - count