# lib/api-zod/src/generated/types/arrowRuleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface ArrowRuleUpdate {
  name?: string;
  numbers?: number[];
  interpretation?: string;
  is_active?: boolean;
}