# lib/api-zod/src/generated/types/missingNumberRuleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface MissingNumberRuleInput {
  missing_number: number;
  title: string;
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
  is_active?: boolean;
}