# lib/api-zod/src/generated/types/numerologyReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */
import type { CalculatedNumbers } from './calculatedNumbers';
import type { LoShuResult } from './loShuResult';
import type { NumerologyReportSubject } from './numerologyReportSubject';
import type { RuleInterpretation } from './ruleInterpretation';

export interface NumerologyReport {
  subject: NumerologyReportSubject;
  numbers: CalculatedNumbers;
  interpretations: RuleInterpretation[];
  lo_shu: LoShuResult;
  generated_at: string;
}