# lib/api-zod/src/generated/types/ruleInterpretation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Numerology CMS API
 * OpenAPI spec version: 0.1.0
 */

export interface RuleInterpretation {
  rule_id: number;
  rule_name: string;
  rule_type: string;
  keywords: string[];
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
  /** @nullable */
  description?: string | null;
}